Utilização de LBA48 para disco (ainda em fase de DEV)

+++ src/Drivers/Disk.h
@@ -1,2 +1,2 @@
-void write_to_sector(short* bytes, unsigned int sector);
+void write_to_sector(short* bytes, unsigned long long sector);
 void read_from_sector(char* buffer, unsigned int sector);
